Optionally read root_dir and windows_root_dir from environment
To enable easy developer/maintainer mode, it is now possible to set
environment variables DLSBUILD_ROOT_DIR and/or DLSBUILD_WIN_ROOT_DIR
and use them to override root_dir and windows_root_dir. A developer will
not have to modify src code in order to do an offline test and generate a
build-server script outside the production environment.