Make python3ext path warning happen only when appropriate.
Before this change, you'd get a warning any time you installed an
executable because the directory you're installing into isn't on the
path. This change avoids the warning when you're installing using a
Pipfile.lock, which should mean that the executable works.

The warning you get is not precisely correct, but I think it's close
enough to help people understand the distinction between installing with
and without a Pipfile.lock.